[ADD] Search bar for BrowseView on iOS 15
